AB#66115 Updating phases and tasks for NCS onboarding
Given an Onboarding Project (PMT), when a Project is created, then the respective Phases and their tasks will be created.

Given the PMT Tasks, when a user is indicating a Status, then they will see the following options:

Not Started
In Progress
Pended
Completed

Given the PMT Tasks, when the tasks are created, then they will default to "Not Started".

Given the PMT Task Status, when a change is made, then that will be tracked in the history.

Given the Phases, when a Contract has wages = true, then it will include the additional Wages Phase & tasks.

Given the PMT Phase, when a user needs to edit multiple tasks, then they will be able to do so.