AB#79947 NCS Implementers Phases and tasks
As a team onboarding new clients, we need our process in sheets to be migrated into Salesforce so that we can better report on the lifecycle of a Client's onboarding.

Given an Implementing Project (PMT), when a Project is created, then the respective Phases and their tasks will be created.
Given the PMT Tasks, when the tasks are created, then they will default to "Not Started".
Given the PMT Task Status, when a change is made, then that will be tracked in the history.